feat: add context for profile data and refactor popular profiles page
