+++ src/pages/profiles/PopularProfiles.jsx
@@ -1,40 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 
 import { Accordion, Container } from "react-bootstrap";
 
-import { axiosReq } from "../../api/axiosDefaults";
 import appStyles from "../../App.module.css";
 import Asset from "../../components/Asset";
-import { useCurrentUser } from "../../contexts/CurrentUserContext";
+import { useProfileData } from "../../contexts/ProfileDataContext";
 import Profile from "./Profile";
 
-// CREDIT: Code Institute Moments Tutorial Project
+// CREDIT: Adapted from Code Institute Moments Tutorial Project
 // URL:    https://github.com/Code-Institute-Solutions/moments
 
 export default function PopularProfiles({ mobile }) {
-  const [profileData, setProfileData] = useState({
-    // pageProfile: { results: [] },
-    popularProfiles: { results: [] },
-  });
-  const { popularProfiles } = profileData;
-  const currentUser = useCurrentUser();
-
-  useEffect(() => {
-    const handleMount = async () => {
-      try {
-        const { data } = await axiosReq.get(
-          "/profiles/?ordering=-followers_count"
-        );
-        setProfileData((prevState) => ({
-          ...prevState,
-          popularProfiles: data,
-        }));
-      } catch (err) {
-        console.log(err);
-      }
-    };
-    handleMount();
-  }, [currentUser]);
+  const { popularProfiles } = useProfileData();
 
   return (
     <Container
